refactor: 로그인,도네이션 모달의 구조 변경
