+++ frontend/src/components/Chat/DonationModal/DonationModal.stories.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import DonationModal from '@/components/DonationModal';
+import DonationModal from './DonationModal';
 
 storiesOf('DonationModal', module).add('default', () => <DonationModal />);
