+++ frontend/src/components/Header/Header.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-
 import { Link } from 'react-router-dom';
 
 import HeaderLogo from '@/assets/images/header-logo.svg';
 import UserIcon from '@/assets/images/user-icon.svg';
 import { flexMixin } from '@/styles/mixins';
+
 import Button from '@/components/Common/Button';
-import LoginModal from '@/components/LoginModal';
+import LoginModal from './LoginModal';
 
 export default function Header({ isSigin }) {
     const [modalOpen, setModalOpen] = useState(false);
